Make frontend use improved read_labels() from core
Since the read_labels() has been updated in the core, we now want to
pass the original untouched data to the read_labels() in the various
tools (tools will process and correctly output invalid data). Update the
LabelInfo interface so that it contains the input file instead of the
list of labels.