Use cache fragments in template
Try to improve response time by using a cache fragment on the repertoire
in the code point listing. Fragment is cached for 3600s (magic number)
and depends on the user and loaded LGR.
For now, cache is invalidated on every save of the LGR. This should be
improved to only be done on repertoire updates.